requirements.txt
Flask==1.1.2
discord.py==1.3.3
python-dotenv==0.13.0
